Move workflow add views to patient shared
