+++ src/js/apps/patients/patient/dashboard/add-workflow_app.js
@@ -4,7 +4,7 @@ import Radio from 'backbone.radio';
 
 import App from 'js/base/app';
 
-import { AddButtonView, i18n, itemClasses } from 'js/views/patients/patient/dashboard/add-workflow_views';
+import { AddButtonView, i18n, itemClasses } from 'js/views/patients/shared/add-workflow/add-workflow_views';
 
 const optEvents = {
   'program-actions': 'add:programAction',
