+++ src/js/i18n/en-US.yml
@@ -377,14 +377,6 @@ careOptsFrontend:
       actionApp:
         notFound: The Action you requested does not exist.
       dashboard:
-        addWorkflowViews:
-          addButtonView:
-            label: Add
-          addWorkflowOptlist:
-            headingText: Add Flow or Action
-            placeholderText: Find...
-          newActionText: New Action...
-          noResultsText: No Published Actions
         dashboardViews:
           emptyView: No Workflows
           addActionBtn: Action
@@ -463,6 +455,16 @@ careOptsFrontend:
           statePatientHeader: State, Patient
           formheader: Form
     shared:
+      addWorkflow:
+        addWorkflowViews:
+          addActionHeadingText: Add Action
+          addButtonView:
+            label: Add
+          addWorkflowOptlist:
+            headingText: Add Flow or Action
+            placeholderText: Find...
+          newActionText: New Action...
+          noResultsText: No Published Actions
       bulkEdit:
         bulkEditViews:
           bulkDeleteActionsSuccess: "{itemCount, plural, one {# Action has been deleted} other {# Actions have been deleted}}"
