+++ src/js/views/patients/shared/add-workflow/add-workflow.scss
